huggingface/requirements.txt
torch
torchvision
fastapi
uvicorn[standard]
pillow
numpy
python-multipart
